Commit 2.8.o

+++ src/net/algem/contact/PersonFileTabView.java
@@ -1,5 +1,5 @@
 /*
- * @(#)PersonFileTabView.java  2.8.m 06/09/13
+ * @(#)PersonFileTabView.java  2.8.o 08/10/13
  *
  * Copyright (c) 1999-2013 Musiques Tangentes All Rights Reserved.
  *
@@ -58,7 +58,7 @@ import net.algem.util.ui.*;
  *
  * @author <a href="mailto:eric@musiques-tangentes.asso.fr">Eric</a>
  * @author <a href="mailto:jmg@musiques-tangentes.asso.fr">Jean-Marc Gobat</a>
- * @version 2.8.m
+ * @version 2.8.o
  */
 public class PersonFileTabView
         extends FileView
@@ -372,8 +372,10 @@ public class PersonFileTabView
   }
   
   void deleteEmployee() {
-    employeeEditor.delete();
-    removeTab(employeeEditor);
+    if (MessagePopup.confirm(this, MessageUtil.getMessage("employee.delete.confirmation"))) { 
+      employeeEditor.delete();
+      removeTab(employeeEditor);
+    }
   }
   
   boolean hasEmployeeChanged() {
